+++ common/darwin/Classes/FlutterWebRTCPlugin.h
@@ -46,6 +46,8 @@ typedef void (^CapturerStopHandler)(CompletionHandler _Nonnull handler);
     NSMutableDictionary<NSString*, RTCFrameCryptor*>* _Nullable frameCryptors;
 @property(nonatomic, strong)
     NSMutableDictionary<NSString*, RTCFrameCryptorKeyProvider*>* _Nullable keyProviders;
+@property(nonatomic, strong)
+    NSMutableDictionary<NSString*, RTCDataPacketCryptor*>* _Nullable dataCryptors;
 
 #if TARGET_OS_IPHONE
 @property(nonatomic, retain)